Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   sparsegpt.py

+++ sparsegpt.py
@@ -247,7 +247,7 @@ class SparseGPT:
                         best_lr = lr
                         best_rho = rho
                         best_max_iter = max_iter
-                        self.layer.weight.data = model.weight.data
+                        self.layer.weight.data = model.weight.data.to(self.layer.weight.data.dtype)
 
         # Print the best hyperparameters and the corresponding loss
         print(f"Best lr: {best_lr}, Best rho: {best_rho}, Best max_iter: {best_max_iter}, Minimum Loss: {min_loss}")
